requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-multipart
python-jose[cryptography]

